Revert to double-precision calculation for interpolation LUTs

+++ src/mixer/ft2_cubic_spline.c
@@ -20,19 +20,19 @@ bool setupCubicSplineTable(void)
 	float *fPtr = fCubicSplineLUT;
 	for (int32_t i = 0; i < CUBIC_SPLINE_PHASES; i++)
 	{
-		const float x1 = (float)i * (1.0f / CUBIC_SPLINE_PHASES);
-		const float x2 = x1 * x1; // x^2
-		const float x3 = x2 * x1; // x^3
+		const double x1 = i * (1.0 / CUBIC_SPLINE_PHASES);
+		const double x2 = x1 * x1; // x^2
+		const double x3 = x2 * x1; // x^3
 
-		const float t1 = (x1 * -0.5f) + (x2 *  1.0f) + (x3 * -0.5f);
-		const float t2 =                (x2 * -2.5f) + (x3 *  1.5f) + 1.0f;
-		const float t3 = (x1 *  0.5f) + (x2 *  2.0f) + (x3 * -1.5f);
-		const float t4 =                (x2 * -0.5f) + (x3 *  0.5f);
+		const double t1 = (x1 * -0.5) + (x2 *  1.0) + (x3 * -0.5);
+		const double t2 =               (x2 * -2.5) + (x3 *  1.5) + 1.0;
+		const double t3 = (x1 *  0.5) + (x2 *  2.0) + (x3 * -1.5);
+		const double t4 =               (x2 * -0.5) + (x3 *  0.5);
 
-		*fPtr++ = t1; // tap #1 at sample offset -1
-		*fPtr++ = t2; // tap #2 at sample offset  0 (center)
-		*fPtr++ = t3; // tap #3 at sample offset  1
-		*fPtr++ = t4; // tap #4 at sample offset  2
+		*fPtr++ = (float)t1; // tap #1 at sample offset -1
+		*fPtr++ = (float)t2; // tap #2 at sample offset  0 (center)
+		*fPtr++ = (float)t3; // tap #3 at sample offset  1
+		*fPtr++ = (float)t4; // tap #4 at sample offset  2
 	}
 
 	/*
@@ -43,23 +43,23 @@ bool setupCubicSplineTable(void)
 	/*
 	for (int32_t i = 0; i < CUBIC_SPLINE_PHASES; i++)
 	{
-		const float x1 = i * (1.0f / CUBIC_SPLINE_PHASES);
-		const float x2 = x1 * x1; // x^2
-		const float x3 = x2 * x1; // x^3
+		const double x1 = i * (1.0 / CUBIC_SPLINE_PHASES);
+		const double x2 = x1 * x1; // x^2
+		const double x3 = x2 * x1; // x^3
 
-		const float t1 = (x1 *  (1.0f/12.0f)) + (x2 * -(1.0f/ 6.0f)) + (x3 *  (1.0f/12.0f));
-		const float t2 = (x1 * -(2.0f/ 3.0f)) + (x2 *  (5.0f/ 4.0f)) + (x3 * -(7.0f/12.0f));
-		const float t3 =                        (x2 * -(7.0f/ 3.0f)) + (x3 *  (4.0f/ 3.0f)) + 1.0f;
-		const float t4 = (x1 *  (2.0f/ 3.0f)) + (x2 *  (5.0f/ 3.0f)) + (x3 * -(4.0f/ 3.0f));
-		const float t5 = (x1 * -(1.0f/12.0f)) + (x2 * -(1.0f/ 2.0f)) + (x3 *  (7.0f/12.0f));
-		const float t6 =                        (x2 *  (1.0f/12.0f)) + (x3 * -(1.0f/12.0f));
+		const double t1 = (x1 *  (1.0/12.0)) + (x2 * -(1.0/ 6.0)) + (x3 *  (1.0/12.0));
+		const double t2 = (x1 * -(2.0/ 3.0)) + (x2 *  (5.0/ 4.0)) + (x3 * -(7.0/12.0));
+		const double t3 =                      (x2 * -(7.0/ 3.0)) + (x3 *  (4.0/ 3.0)) + 1.0;
+		const double t4 = (x1 *  (2.0/ 3.0)) + (x2 *  (5.0/ 3.0)) + (x3 * -(4.0/ 3.0));
+		const double t5 = (x1 * -(1.0/12.0)) + (x2 * -(1.0/ 2.0)) + (x3 *  (7.0/12.0));
+		const double t6 =                      (x2 *  (1.0/12.0)) + (x3 * -(1.0/12.0));
 
-		*fPtr++ = t1; // tap #1 at sample offset -2
-		*fPtr++ = t2; // tap #2 at sample offset -1
-		*fPtr++ = t3; // tap #3 at sample offset  0 (center)
-		*fPtr++ = t4; // tap #4 at sample offset  1
-		*fPtr++ = t5; // tap #5 at sample offset  2
-		*fPtr++ = t6; // tap #6 at sample offset  3
+		*fPtr++ = (float)t1; // tap #1 at sample offset -2
+		*fPtr++ = (float)t2; // tap #2 at sample offset -1
+		*fPtr++ = (float)t3; // tap #3 at sample offset  0 (center)
+		*fPtr++ = (float)t4; // tap #4 at sample offset  1
+		*fPtr++ = (float)t5; // tap #5 at sample offset  2
+		*fPtr++ = (float)t6; // tap #6 at sample offset  3
 	}
 	*/
